+++ example/lib/random_map/map_generator.dart
@@ -67,7 +67,7 @@ class MapGenerator {
       //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
       //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
       // ],
-      builder: _buildTerrainBuilder().build,
+      builder: _buildTerrainBuilder().builder,
     );
   }