+++ lib/map/base/tile.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/map/map_paint.dart';
+import 'package:bonfire/map/util/map_paint.dart';
 import 'package:bonfire/util/controlled_update_animation.dart';
 import 'package:flutter/widgets.dart';